s/int/long memory view

+++ skimage/transform/_hough_transform.pyx
@@ -149,7 +149,7 @@ def hough_ellipse(cnp.ndarray img, int threshold=4, double accuracy=1,
     if img.ndim != 2:
             raise ValueError('The input image must be 2D.')
 
-    cdef int[:, :] pixels = np.transpose(np.nonzero(img))
+    cdef long[:, :] pixels = np.transpose(np.nonzero(img))
     cdef Py_ssize_t num_pixels = pixels.shape[0]
     cdef list acc = list()
     cdef list results = list()
